python_service/requirements.txt
grpcio
grpcio-tools
msgpack
zstandard
numpy
pandas
scikit-learn